+++ example/lib/main.dart
@@ -1,5 +1,6 @@
-import 'package:example/pages/steven_example.dart';
 import 'package:flutter/material.dart';
+import 'package:simple_animated_rating_bar/simple_animated_rating_bar.dart';
+import 'package:simple_animated_rating_bar/utils/arb_type.dart';
 
 void main() {
   runApp(
@@ -14,7 +15,26 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const StevenUniverseExample();
-    // return DefaultExample();
+    return Center(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          AnimatedRatingBar(
+            emptyWidget: Container(
+              height: 50,
+              width: 50,
+              color: Colors.red,
+            ),
+            fullWidget: Container(
+              height: 50,
+              width: 50,
+              color: Colors.green,
+            ),
+            ratingLength: 6,
+            animationType: ARBAnimationType.rotate,
+          )
+        ],
+      ),
+    );
   }
 }